% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{com_extract}
\alias{com_extract}
\title{Compute the custom hyperbolic after number extraction}
\usage{
com_extract(vect)
}
\arguments{
\item{vect}{A character vector}
}
\value{
A numeric vector of length one
}
\description{
Compute the custom hyperbolic after number extraction
}
