% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{num_extraction}
\alias{num_extraction}
\title{Retrieving numbers from a character}
\usage{
num_extraction(vect)
}
\arguments{
\item{vect}{A character vector}
}
\value{
A numeric vector without the non digits
}
\description{
Given a vector, this function remove all non digits characters
}
